refactor(l5): move graph-derivation from src/ into server/domain (issue 864)
Moves the graph-derivation primitive out of src/ into server/domain, splitting
the single module into the model, the persistence statements, and the
orchestration entry point, and bringing the moved tests to the repo lint bar
under the #878 transformations.

Sources, old -> new:

  src/graph-derivation.ts -> server/domain/graph-derivation.ts
                             server/domain/graph-derivation-model.ts
                             server/domain/graph-derivation-persistence.ts

Tests, old -> new. The 24 cases in the single src/ file are redistributed by
describe subject into five siblings, names and assertion meaning unchanged:

  src/graph-derivation.test.ts (24 cases)
    -> server/domain/graph-derivation-lifecycle.test.ts       (5)
    -> server/domain/graph-derivation-anchor-identity.test.ts (5)
    -> server/domain/graph-derivation-prune.test.ts           (2)
    -> server/domain/graph-derivation-isolation.test.ts       (6)
    -> server/domain/graph-derivation-content-free.test.ts    (6)

  src/graph-derivation.live.test.ts
    -> server/domain/graph-derivation.live.test.ts
  shared harness
    -> server/domain/graph-derivation-test-helpers.ts

The #878 transformations applied to the moved tests: non-null assertions
replaced with expectDefined from scripts/test-support/expect-defined.ts; every
it()/describe() body over the function rule hoisted to a module-scope function
called by the same name; `as any` replaced with `as unknown as <type>`; unused
imports removed. The FakeGraph query dispatcher is split into one private
handler per SQL shape so each stays inside the function-size and complexity
rules, and the fake's result rows are typed instead of cast per return.

src/graph-derivation.ts stays as an L5 re-export shim; it is retired with src/
at L6. Shim closure moves 62 -> 61 counting shims as excluded.

Verified: oxlint --deny-warnings clean on every touched file; bunx tsc --noEmit
exit 0; bun run test:isolated over the five split files 24 pass / 0 fail,
matching the 24 cases at origin/main; scripts/done-means/864-moved-out-of-src.sh
PASS.
